// web-application/src/types-openapi/api/models/CreateVoucherRequest.ts
export interface CreateVoucherRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateVoucherRequest
     */
    code?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateVoucherRequest
     */
    title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateVoucherRequest
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateVoucherRequest
     */
    discountType?: string | null;
    /**
     * 
     * @type {number}
     * @memberof CreateVoucherRequest
     */
    discountValue?: number;
    /**
     * 
     * @type {number}
     * @memberof CreateVoucherRequest
     */
    discountPercentage?: number | null;
    /**
     * 
     * @type {number}
     * @memberof CreateVoucherRequest
     */
    maxDiscountValue?: number | null;
    /**
     * 
     * @type {number}
     * @memberof CreateVoucherRequest
     */
    minOrderValue?: number | null;
    /**
     * 
     * @type {Date}
     * @memberof CreateVoucherRequest
     */
    startAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof CreateVoucherRequest
     */
    endAt?: Date;
    /**
     * 
     * @type {number}
     * @memberof CreateVoucherRequest
     */
    usageLimitTotal?: number;
    /**
     * 
     * @type {number}
     * @memberof CreateVoucherRequest
     */
    usageLimitPerUser?: number;
    /**
     * 
     * @type {boolean}
     * @memberof CreateVoucherRequest
     */
    isActive?: boolean;
    /**
     * 
     * @type {Array<VoucherTimeRuleDto>}
     * @memberof CreateVoucherRequest
     */
    timeRules?: Array<VoucherTimeRuleDto> | null;
    /**
     * 
     * @type {Array<VoucherUserRuleDto>}
     * @memberof CreateVoucherRequest
     */
    userRules?: Array<VoucherUserRuleDto> | null;
}